requirements.txt
Python==3.8.10
Django==4.2
djangorestframework==3.14.0